+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/edit/EditToolbar.kt
@@ -14,6 +14,7 @@ import android.view.inputmethod.InputMethodManager
 import android.widget.ImageView
 import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.browser.toolbar.R
+import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.support.ktx.android.content.res.pxToDp
 import mozilla.components.support.ktx.android.view.showKeyboard
 import mozilla.components.ui.autocomplete.InlineAutocompleteEditText
@@ -48,6 +49,9 @@ class EditToolbar(
         setSelectAllOnFocus(true)
 
         setOnCommitListener { toolbar.onUrlEntered(text.toString()) }
+        setOnTextChangeListener { text, _ ->
+            editListener?.onTextChanged(text)
+        }
     }
 
     private val cancelView = ImageView(context).apply {
@@ -60,6 +64,8 @@ class EditToolbar(
             toolbar.displayMode()
         }
     }
+
+    internal var editListener: Toolbar.OnEditListener? = null
 
     init {
         addView(urlView)