+++ mobile/android/android-components/components/concept/toolbar/src/main/java/mozilla/components/concept/toolbar/Toolbar.kt
@@ -76,6 +76,31 @@ interface Toolbar {
     fun asView(): View = this as View
 
     /**
+     * Registers the given listener to be invoked when the user edits the URL.
+     */
+    fun setOnEditListener(listener: OnEditListener)
+
+    /**
+     * Listener to be invoked when the user edits the URL.
+     */
+    interface OnEditListener {
+        /**
+         * Fired when the toolbar switches to edit mode.
+         */
+        fun onStartEditing() = Unit
+
+        /**
+         * Fired when the toolbar switches back to display mode.
+         */
+        fun onStopEditing() = Unit
+
+        /**
+         * Fired whenever the user changes the text in the address bar.
+         */
+        fun onTextChanged(text: String) = Unit
+    }
+
+    /**
      * Generic interface for actions to be added to the toolbar.
      */
     interface Action {